Better organise code. Fix failing test cases
