+++ src/main/config.ts
@@ -1,7 +1,7 @@
 /* eslint-disable import/prefer-default-export */
 
-import { TranscriptionEngine } from './handlers/media/transcriptionEngines';
+import { TranscriptionEngine } from './handlers/media/transcribeTypes';
 
-// Whether to use the sample transcription instead of actually transcribing.
-// Useful for testing.
+// Which transcription engine to use for any transcription.
+// This will eventually be replaced with a config file allowing the user to choose which engine they want
 export const TRANSCRIPTION_ENGINE = TranscriptionEngine.DEEPSPEECH;
